+++ Assets/Game/Scripts/Services/Chess/VOs/HintVO.cs
@@ -21,7 +21,7 @@ namespace TurboLabz.Chess
         public ChessSquare fromSquare;
         public ChessSquare toSquare;
         public bool isHindsight;
-        public int strength;
+        public float strength;
         public string piece;
         public string skinId;
         public bool didPlayerMadeBestMove;
